API and JSP

+++ src/es/upm/dit/geoloc/dao/ThoughtDAO.java
@@ -5,7 +5,7 @@ import es.upm.dit.geoloc.dao.model.Thought;
 public interface ThoughtDAO {
 	
 	public Integer createThought(Thought thought);
-	public Thought readThought(double id);
+	public Thought readThought(int id);
 	public void updateThought(Thought thought);
 	public void deleteThought(Thought thought);
